Significant Progress on the moving around
