+++ src/PeepingTom/Fast.hs
@@ -1,0 +1,3 @@
+module PeepingTom.Fast where
+
+import PeepingTom.Fast.Scan (ScanOptions (..), defaultScanOptions, scanMap, scanMapS, updateState, updateStateS)
