+++ src/PeepingTom/Internal.hs
@@ -1,7 +1,5 @@
 module PeepingTom.Internal where
 
-import Text.Read
-
 type PID = Int
 type FD = Int
 type Address = Int
